pyproject.toml
[project]
name = "ls-py175-flash-cards"
version = "0.1.0"
description = ""
authors = [
    {name = "Justin Schultz",email = "SchultzJustin@ymail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "ls_py175_flash_cards", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
